Module randomx::ffi

Structs

__darwin_pthread_handler_rec

__va_list_tag

_opaque_pthread_attr_t

_opaque_pthread_cond_t

_opaque_pthread_condattr_t

_opaque_pthread_mutex_t

_opaque_pthread_mutexattr_t

_opaque_pthread_once_t

_opaque_pthread_rwlock_t

_opaque_pthread_rwlockattr_t

_opaque_pthread_t

randomx_cache

randomx_dataset

randomx_vm

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INT64_MAX

INT64_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST64_MAX

INT_FAST64_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST64_MAX

INT_LEAST64_MIN

RANDOMX_DATASET_ITEM_SIZE

RANDOMX_HASH_SIZE

RSIZE_MAX

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINT64_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_FAST64_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

UINT_LEAST64_MAX

WINT_MAX

WINT_MIN

_DARWIN_FEATURE_64_BIT_INODE

_DARWIN_FEATURE_ONLY_UNIX_CONFORMANCE

_DARWIN_FEATURE_UNIX_CONFORMANCE

_LIBCPP_HAS_MERGED_TYPEINFO_NAMES_DEFAULT

__DARWIN_64_BIT_INO_T

__DARWIN_C_ANSI

__DARWIN_C_FULL

__DARWIN_C_LEVEL

__DARWIN_NON_CANCELABLE

__DARWIN_NO_LONG_LONG

__DARWIN_ONLY_64_BIT_INO_T

__DARWIN_ONLY_UNIX_CONFORMANCE

__DARWIN_ONLY_VERS_1050

__DARWIN_SUF_64_BIT_INO_T

__DARWIN_SUF_1050

__DARWIN_SUF_EXTSN

__DARWIN_UNIX03

__DARWIN_VERS_1050

__PTHREAD_ATTR_SIZE__

__PTHREAD_CONDATTR_SIZE__

__PTHREAD_COND_SIZE__

__PTHREAD_MUTEXATTR_SIZE__

__PTHREAD_MUTEX_SIZE__

__PTHREAD_ONCE_SIZE__

__PTHREAD_RWLOCKATTR_SIZE__

__PTHREAD_RWLOCK_SIZE__

__PTHREAD_SIZE__

__STDC_WANT_LIB_EXT1__

__WORDSIZE

randomx_flags_RANDOMX_FLAG_DEFAULT

randomx_flags_RANDOMX_FLAG_FULL_MEM

randomx_flags_RANDOMX_FLAG_HARD_AES

randomx_flags_RANDOMX_FLAG_JIT

randomx_flags_RANDOMX_FLAG_LARGE_PAGES

Functions

randomx_alloc_cache⚠

Creates a randomx_cache structure and allocates memory for RandomX Cache.

randomx_alloc_dataset⚠

Creates a randomx_dataset structure and allocates memory for RandomX Dataset.

randomx_calculate_hash⚠

Calculates a RandomX hash value.

randomx_create_vm⚠

Creates and initializes a RandomX virtual machine.

randomx_dataset_item_count⚠

Gets the number of items contained in the dataset.

randomx_destroy_vm⚠

Releases all memory occupied by the randomx_vm structure.

randomx_get_dataset_memory⚠

Returns a pointer to the internal memory buffer of the dataset structure. The size of the internal memory buffer is randomx_dataset_item_count() * RANDOMX_DATASET_ITEM_SIZE.

randomx_init_cache⚠

Initializes the cache memory and SuperscalarHash using the provided key value.

randomx_init_dataset⚠

Initializes dataset items.

randomx_release_cache⚠

Releases all memory occupied by the randomx_cache structure.

randomx_release_dataset⚠

Releases all memory occupied by the randomx_dataset structure.

randomx_vm_set_cache⚠

Reinitializes a virtual machine with a new Cache. This function should be called anytime the Cache is reinitialized with a new key.

randomx_vm_set_dataset⚠

Reinitializes a virtual machine with a new Dataset.

Type Definitions

__builtin_va_list

__darwin_blkcnt_t

__darwin_blksize_t

__darwin_clock_t

__darwin_ct_rune_t

__darwin_dev_t

__darwin_fsblkcnt_t

__darwin_fsfilcnt_t

__darwin_gid_t

__darwin_id_t

__darwin_ino64_t

__darwin_ino_t

__darwin_intptr_t

__darwin_mach_port_name_t

__darwin_mach_port_t

__darwin_mbstate_t

__darwin_mode_t

__darwin_natural_t

__darwin_off_t

__darwin_pid_t

__darwin_pthread_attr_t

__darwin_pthread_cond_t

__darwin_pthread_condattr_t

__darwin_pthread_key_t

__darwin_pthread_mutex_t

__darwin_pthread_mutexattr_t

__darwin_pthread_once_t

__darwin_pthread_rwlock_t

__darwin_pthread_rwlockattr_t

__darwin_pthread_t

__darwin_ptrdiff_t

__darwin_rune_t

__darwin_sigset_t

__darwin_size_t

__darwin_socklen_t

__darwin_ssize_t

__darwin_suseconds_t

__darwin_time_t

__darwin_uid_t

__darwin_useconds_t

__darwin_uuid_string_t

__darwin_uuid_t

__darwin_va_list

__darwin_wchar_t

__darwin_wint_t

__int8_t

__int16_t

__int32_t

__int64_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

randomx_flags

register_t

syscall_arg_t

u_int8_t

u_int16_t

u_int32_t

u_int64_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

user_addr_t

user_long_t

user_off_t

user_size_t

user_ssize_t

user_time_t

user_ulong_t

wchar_t

Unions

__mbstate_t